requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docutils==0.21.2
feedgenerator==2.1.0
feedparser==6.0.11
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
newspaper4k==0.9.3.1
nltk==3.8.1
numpy==2.2.1
openai==1.59.8
ordered-set==4.1.0
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tldextract==5.1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
watchfiles==1.0.4
